requirements.txt
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-celery-beat==2.2.1
django-graphql-jwt==0.3.4
django-timezone-field==4.2.1
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
kombu==5.1.0
promise==2.3
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
python-decouple==3.6